requirements.txt
Flask
Flask-SQLAlchemy
Werkzeug
psycopg2-binary
python-dotenv
pymysql
rpm
npm
